Major restructure of main class and project
This involes several major changes:
 * "Captures" were created for modularity
    - now video file/live camera captures are implementations of a
common interface
 * Spinnaker SDK libraries can optionally be compiled
    - now the "no-camera" in the windows makefile can be used to not
include Spinnaker SDK libraries/headers
 * Work on creating components of the final structure started
