+++ src/Capture/Capture.h
@@ -1,0 +1,12 @@
+#include <opencv2/core/core.hpp>
+#include <opencv2/highgui/highgui.hpp>
+#include <opencv2/imgproc/imgproc.hpp>
+#include <opencv2/videoio.hpp>
+#include <chrono> 
+
+// simple interface to handle captures nicely
+class Capture{
+public:
+    virtual cv::Mat* nextFrame() = 0;
+    double waittime; // time until next frame: 0 for live footage, fps for file
+};
